+++ frontend/src/App.tsx
@@ -15,6 +15,7 @@ import RequireAuth from './components/require-auth';
 import useThemeClass from './utils/theme-changer';
 import AdminContainer from './components/admin-container';
 import { Role } from './types/role-type.tsx';
+import EditBlog from './pages/edit-blog.tsx';
 
 function App() {
   useLayoutEffect(() => {
@@ -34,6 +35,7 @@ function App() {
             </Route>
             <Route element={<RequireAuth allowedRole={[Role.Admin, Role.User]} />}>
               <Route path="add-blog" element={<AddBlog />} />
+              <Route path="edit-blog/:postId" element={<EditBlog />} />
             </Route>
             <Route path="admin" element={<RequireAuth allowedRole={[Role.Admin]} />}>
               <Route element={<AdminContainer />}>
