+++ frontend/src/types/post-type.tsx
@@ -6,6 +6,7 @@ type Post = {
   timeOfPost: string;
   description: string;
   categories: string[];
+  authorId?: string;
 };
 
 export default Post;
